Split out `canvasToBlob`
